bot/requirements.txt
python-telegram-bot>=20.0
python-dotenv>=0.19.2
stripe>=2.60.0  
yookassa>=3.0.0  
python-dotenv==1.0.0
psycopg2-binary==2.9.3  
SQLAlchemy==2.0.0
openai==0.27.2
requests==2.31.0  
gunicorn==20.1.0 
pytest==7.4.0
pytest-mock==3.10.0
pytest-asyncio>=0.21.0 
aiohttp>=3.8.0  
python-dateutil>=2.8.2  
typing-extensions>=4.0.0 